return null;
at the end of the two main functions, add "return null;" because
otherwise Emacs javascript parser whinges about not returning something
from a routine.  I guess it is in the standard and all that.

apologies if this is noise, but that's why I'm making tiny commits...you
can ignore them.